# Optimize FBGEMM Triton MX4 Quantize